// src/app/theme/models/assessment.model.ts
export class AssessmentItem {
    id: number;
    projectId: number;
    companyId: number;
    questionId: number;
    answer: string;
    excludeInReporting: string;
    attachementAvailable: string;
    attachementData: string;
    status: string;
    comments: string;
    question: string;
    questionType: string;
}